app/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-money==1.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
psycopg2-binary==2.8.4
py-moneyed==0.8.0
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
